undo float32

+++ morphax/data.py
@@ -92,7 +92,7 @@ def image_to_jnp(files_path):
             dat = img
         else:
             dat = jnp.append(dat,img,0)
-    return dat.astype(jnp.float32)
+    return dat
 
 #Save images
 def save_images(images,files_path):
